# data/swagger.yaml
swagger: '2.0'
info:
  description: to be written
  version: 0.1.0
  title: Hourglass
  termsOfService: ''
  contact:
    name: busoc ground controller
    email: busocgc@busoc.be
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'virtualtest.busoc.be:9000'
basePath: /
consumes:
  - application/json
produces:
  - application/json
  - application/x-gzip
securityDefinitions:
  api_key:
    type: apiKey
    description: json web token
    in: header
    name: Authorization
  basicAuth:
    type: basic
...
schemes:
  - http
...
paths:
  /auth:
    post:
      tags:
        - Authentication
        - Users
      summary: authenticate user
      operationId: Signin
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - user
              - passwd
            properties:
              user:
                type: string
                example: rla
              passwd:
                type: string
                example: helloworld
      responses:
        '200':
          description: authentication succeed
        '400':
          description: invalid input provided
        '401':
          description: invalid credentials provided
  /users/:
    post:
      tags:
        - Users
      summary: register user
      operationId: RegisterUser
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              firstname:
                type: string
                description: user's first name
                example: roger
              lastname:
                type: string
                description: user's last name
                example: lamotte
              initial:
                type: string
                description: user's initial
                example: rla
                minLength: 2
                maxLength: 3
              internal:
                type: boolean
              email:
                type: string
              passwd:
                type: string
              positions:
                type: array
                items:
                  type: string
      responses:
        '201':
          description: user registered
          schema:
            $ref: '#/definitions/User'
        '401':
          description: unauthenticated user
    get:
      tags:
        - Users
      summary: list of users
      operationId: ListUsers
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: list users
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '204':
          description: no user available
        '401':
          description: unauthenticated user
  '/users/{id}':
    parameters:
      - name: id
        in: path
        description: user id
        required: true
        type: integer
        format: int64
    get:
      tags:
        - Users
      description: view user detail
      operationId: ViewUser
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: operation succeed
          schema:
            $ref: '#/definitions/User'
        '401':
          description: unauthenticated user
        '404':
          description: user not found
    put:
      tags:
        - Users
      description: update user information and settings
      operationId: UpdateUser
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: operation succeed
          schema:
            $ref: '#/definitions/User'
        '401':
          description: unauthenticated user
        '404':
          description: user not found
  '/users/{id}/passwd':
    parameters:
      - name: id
        in: path
        description: user id
        required: true
        type: integer
        format: int64
    put:
      tags:
        - Users
      summary: update user password
      operationId: UpdatePasswd
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '204':
          description: password updated
        '401':
          description: unauthenticated user
        '403':
          description: user try to update password of another user
  /categories/:
    get:
      tags:
        - Categories
      summary: list of categories
      operationId: ListCategories
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: list categories
          schema:
            type: array
            items:
              $ref: '#/definitions/Category'
        '204':
          description: no categories available
        '401':
          description: unauthenticated user
    post:
      tags:
        - Categories
      summary: create new category
      operationId: NewCategory
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Category'
      responses:
        '201':
          description: category successfully created
          schema:
            $ref: '#/definitions/Category'
        '400':
          description: invalid input provided
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
  '/categories/{id}':
    parameters:
      - name: id
        in: path
        description: category id
        required: true
        type: integer
        format: int64
    get:
      tags:
        - Categories
      summary: get detailed information about a category
      operationId: ViewCategory
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: operation succeed
        '400':
          description: bad request
        '404':
          description: not found
    put:
      tags:
        - Categories
      summary: update category information
      operationId: UpdateCategory
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: operation succeed
          schema:
            $ref: '#/definitions/Category'
        '400':
          description: bad request
        '404':
          description: not found
  /transfers/:
    get:
      tags:
        - Transfers
      summary: list of transfers
      operationId: ListTransfers
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - in: query
          name: dtstart
          type: string
          format: date-time
        - in: query
          name: dtend
          type: string
          format: date-time
        - in: query
          name: 'category[]'
          collectionFormat: multi
          type: array
          items:
            type: string
        - in: query
          name: 'status[]'
          collectionFormat: multi
          type: array
          items:
            type: string
      responses:
        '200':
          description: list of transfers (scheduled to completed)
          schema:
            $ref: '#/definitions/Transfer'
        '204':
          description: no transfers available
        '401':
          description: unauthenticated user
    post:
      tags:
        - Transfers
      summary: create a transfer
      operationId: NewTransfer
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - name: body
          in: body
          schema:
            type: object
            required:
              - uplink
              - event
            properties:
              uplink:
                type: integer
                format: int64
                example: 34
              event:
                type: integer
                format: int64
                example: 563
      responses:
        '201':
          description: transfer successfully created
          schema:
            $ref: '#/definitions/Transfer'
        '400':
          description: invalid input provided
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested action
  '/transfers/{id}':
    parameters:
      - name: id
        in: path
        description: transfer id
        required: true
        type: integer
        format: int64
    get:
      tags:
        - Transfers
      summary: get detailed information about transfer
      operationId: ViewTransfer
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: operation succeed
          schema:
            $ref: '#/definitions/Transfer'
        '401':
          description: unauthenticated user
        '404':
          description: uplink not found
    put:
      tags:
        - Transfers
      summary: update transfer status
      operationId: UpdateTransfer
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Status'
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: transfer successfully updated
          schema:
            $ref: '#/definitions/Transfer'
        '400':
          description: invalid input provided
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
  /uplinks/:
    get:
      tags:
        - Uplinks
      summary: list of scheduled uplinks
      operationId: ListUplinks
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - in: query
          name: dtstart
          type: string
          format: date-time
        - in: query
          name: dtend
          type: string
          format: date-time
        - in: query
          name: 'category[]'
          collectionFormat: multi
          type: array
          items:
            type: string
        - in: query
          name: 'status[]'
          collectionFormat: multi
          type: array
          items:
            type: string
      responses:
        '200':
          description: list of uplinks (scheduled to completed)
          schema:
            $ref: '#/definitions/Uplink'
        '204':
          description: no uplinks available
        '401':
          description: unauthenticated user
    post:
      tags:
        - Uplinks
      summary: schedule a new uplink
      operationId: NewUplink
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - event
              - slot
              - file
            properties:
              event:
                type: integer
                format: int64
                example: 3139293
              slot:
                type: integer
                format: int64
                example: 34635621
              file:
                type: integer
                format: int64
                example: 219
      responses:
        '201':
          description: uplink successfully created
          schema:
            $ref: '#/definitions/Uplink'
        '400':
          description: invalid input provided
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
  '/uplinks/{id}':
    parameters:
      - name: id
        in: path
        description: uplink id
        required: true
        type: integer
        format: int64
    get:
      tags:
        - Uplinks
      summary: get detailed information about uplink
      operationId: ViewUplink
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: operation succeed
          schema:
            $ref: '#/definitions/Uplink'
        '401':
          description: unauthenticated user
        '404':
          description: uplink not found
    put:
      tags:
        - Uplinks
      summary: update uplink status
      operationId: UpdateUplink
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Status'
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: uplink successfully updated
          schema:
            $ref: '#/definitions/Uplink'
        '400':
          description: invalid input provided
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
  /downlinks/:
    get:
      tags:
        - Downlinks
      summary: List of scheduled downlinks
      operationId: ListDownlinks
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - in: query
          name: dtstart
          type: string
          format: date-time
        - in: query
          name: dtend
          type: string
          format: date-time
        - in: query
          name: 'category[]'
          collectionFormat: multi
          type: array
          items:
            type: string
        - in: query
          name: 'status[]'
          collectionFormat: multi
          type: array
          items:
            type: string
      responses:
        '200':
          description: list downlinks
          schema:
            type: array
            items:
              $ref: '#/definitions/Downlink'
        '204':
          description: no downlinks available
        '401':
          description: unauthenticated user
    post:
      tags:
        - Downlinks
      summary: create new downlink
      operationId: NewDownlink
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - event
              - slot
              - file
            properties:
              event:
                type: integer
                format: int64
                example: 10354
              slot:
                type: integer
                format: int64
                example: 34635621
              file:
                type: integer
                format: int64
                example: 2
      responses:
        '201':
          description: downlink successfully created
          schema:
            $ref: '#/definitions/Downlink'
        '400':
          description: invalid input provided
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
  '/downlinks/{id}':
    parameters:
      - name: id
        in: path
        description: downlink id
        required: true
        type: integer
        format: int64
    get:
      tags:
        - Downlinks
      summary: get detailed information about downlink
      operationId: ViewDownlink
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: operation succeed
          schema:
            $ref: '#/definitions/Downlink'
        '401':
          description: unauthenticated user
        '404':
          description: uplink not found
    put:
      tags:
        - Downlinks
      summary: update downlink status
      operationId: UpdateDownlink
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Status'
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: downlink successfully updated
          schema:
            $ref: '#/definitions/Downlink'
        '400':
          description: invalid input provided
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
  /slots/:
    get:
      tags:
        - Slots
      summary: list of slots
      operationId: ListSlots
      parameters:
        - name: 'category[]'
          in: query
          collectionFormat: multi
          required: false
          type: array
          items:
            type: string
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: list of slots
        '204':
          description: no slots available
        '401':
          description: unauthenticated user
    post:
      tags:
        - Slots
      operationId: NewSlot
      summary: register a new slot
      security:
        - basicAuth: []
        - api_key: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Slot'
      responses:
        '201':
          description: slot successfully created
          schema:
            $ref: '#/definitions/Slot'
        '400':
          description: invalid input provided
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
  '/slots/{id}':
    parameters:
      - name: id
        in: path
        description: slot id
        required: true
        type: integer
        format: int64
    get:
      tags:
        - Slots
      summary: get detailed information about a slot
      operationId: ViewSlot
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: operation succeed
          schema:
            $ref: '#/definitions/Slot'
        '401':
          description: unauthenticated user
        '404':
          description: slot not found
    delete:
      tags:
        - Slots
      summary: delete a slot
      operationId: DeleteSlot
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '204':
          description: operation succeed
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
        '404':
          description: slot not found
  /files/:
    get:
      tags:
        - Files
      summary: list of files
      operationId: ListFiles
      parameters:
        - name: 'category[]'
          required: false
          collectionFormat: multi
          in: query
          type: array
          items:
            type: string
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: list of files
          schema:
            type: array
            items:
              $ref: '#/definitions/File'
        '204':
          description: no files available
        '401':
          description: unauthenticated user
    post:
      tags:
        - Files
      summary: create a new file
      operationId: NewFile
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/File'
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '201':
          description: file successfully created
          schema:
            $ref: '#/definitions/File'
        '400':
          description: invalid input provided
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
...
  /todos/:
    get:
      tags:
        - Todos
      summary: list of actions/todos
      description: list of actions/todos
      operationId: ListTodos
      parameters:
        - name: 'category[]'
          in: query
          required: false
          collectionFormat: multi
          type: array
          items:
            type: string
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '200':
          description: list of actions/todos
          schema:
            type: array
            items:
              $ref: '#/definitions/Todo'
        '204':
          description: no todos available
        '401':
          description: unauthenticated user
    post:
      tags:
        - Todos
      summary: create a todo/action
      description: create a new todo/action
      operationId: NewTodo
      security:
        - basicAuth: []
        - api_key: []
      responses:
        '201':
          description: operation succeed
          schema:
            $ref: '#/definitions/Todo'
        '400':
          description: invalid input
        '401':
          description: unauthenticated user
        '403':
          description: authenticated user is no allowed to perform requested actions
...
  '/events/{source}/':
    post:
      tags:
        - Events
      summary: import events
      operationId: ImportEvents
      description: ''
      security:
        - basicAuth: []
        - api_key: []
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: source
          in: path
          description: source of events
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              source:
                type: string
                example: optimis
              events:
                type: array
                items:
                  $ref: '#/definitions/Event'
      responses:
        '204':
          description: operation succeed
        '400':
          description: bad request
...
definitions:
  Status:
    description: status
    type: string
    enum:
      - n/a
      - tentative
      - scheduled
      - on going
      - completed
      - canceled
      - aborted
...
  Category:
    type: object
    required:
      - name
    properties:
      uid:
        type: integer
        format: int64
        readOnly: true
        example: 372
      name:
        type: string
        example: payload
      categories:
        type: array
        items:
          type: string
      lastmod:
        type: string
        format: date-time
        example: '2017-09-12T34:45:00Z'
        readOnly: true
      user:
        type: string
        readOnly: true
        example: nbr
...
  User:
    type: object
    properties:
      uid:
        type: integer
        format: int64
        example: 2
      firstname:
        type: string
        example: roger
      lastname:
        type: string
        example: lamotte
      initial:
        type: string
        minLength: 3
        example: rla
      email:
        type: string
        format: email
      internal:
        type: boolean
        default: false
      settings:
        type: object
      positions:
        type: array
        items:
          type: string
          example:
            - gc
            - dev
...